[#35] 가격에 대한 정보를 포함한 roomResponse를 주기 위해 이름을 수정함
가격정보 데이터를 포함하여 응답하기 위해서 기존 RoomResponse 클래스를 RoomInfo로 수정함
RoomResponse는 가격과 숙소 정보를 모두 포함한 dto로 만듦
